requirements.txt
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
mailchimp-marketing==3.0.80
MarkupSafe==3.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.4
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.3
